refactor(coordination): move CoordinationService to base.py; keep __init__ thin
Addresses review feedback: no implementation in __init__.py. CoordinationService now
lives in coordination/base.py (matching the base.py convention used elsewhere, e.g.
commands/distributed_lock/base.py); __init__ only holds the package docstring and
re-exports (CoordinationService, SignalListener, CoordinationBackendUnavailableError),
so `from superset.coordination import ...` is unchanged.
